Change HTML interpolations to getters
We were seeing lots of ctx.post[0] is undefined errors in the console because post[0] was undefined before the observables resolved. Using a getter function allows us to use a ternary operator to return null rather than undefined and catch the error while we wait for the observable to resolve.