implements first attempt at pinch zoom
Can't find an easy way to test this on mobile because ngrok doesn't work due to the strapi backend being hosted on a different localhost port so just going to try pushing to live website and testing it there. It shouldn't break anything else so should be safe to do so.